+++ frontend/src/Logs.jsx
@@ -5,7 +5,7 @@ import Grid from "@mui/material/Grid";
 import MenuItem from "@mui/material/MenuItem";
 import Button from "@mui/material/Button";
 import Divider from "@mui/material/Divider";
-import { Link, useNavigate, useParams, useLocation } from 'react-router-dom';
+import { Link, useNavigate, useParams, useLocation } from 'react-router';
 import DownloadIcon from '@mui/icons-material/Download';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';