+++ frontend/src/SingleCalibrationPage.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { useParams,  useNavigate,  Link } from "react-router-dom";
+import { useParams,  useNavigate,  Link } from "react-router";
 import { useConfirm } from 'material-ui-confirm';
 import { CircularProgress, Button, Typography, Box } from "@mui/material";
 import Dialog from '@mui/material/Dialog';