+++ frontend/src/StartNewExperiment.jsx
@@ -11,7 +11,7 @@ import {Typography} from '@mui/material';
 import Button from "@mui/material/Button";
 import TextField from '@mui/material/TextField';
 import Autocomplete, { createFilterOptions } from '@mui/material/Autocomplete';
-import {useNavigate } from 'react-router-dom';
+import {useNavigate } from 'react-router';
 import SaveIcon from '@mui/icons-material/Save';
 import LoadingButton from '@mui/lab/LoadingButton';