+++ frontend/src/components/DisplayProfile.jsx
@@ -8,7 +8,7 @@ import UnderlineSpan from "./UnderlineSpan";
 import CalculateOutlinedIcon from '@mui/icons-material/CalculateOutlined';
 import PioreactorIcon from "./PioreactorIcon"
 import PioreactorsIcon from './PioreactorsIcon';
-import { Link } from 'react-router-dom';
+import { Link } from 'react-router';
 import ViewTimelineOutlinedIcon from '@mui/icons-material/ViewTimelineOutlined';