+++ frontend/src/components/ManageInventoryMenu.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import MenuItem from '@mui/material/MenuItem';
 import Button from '@mui/material/Button';
 import { useConfirm } from 'material-ui-confirm';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router';
 import Menu from "@mui/material/Menu";
 import ListItemText from "@mui/material/ListItemText";
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';